TUP Canteen POS
A project we did during my second-year college. A POS for the stalls inside the campus. It has an admin domain to keep track of the current stalls, a staff domain to make an order and to keep track of the current stocks they have left. Lastly, the program also generates a receipt every purchase.